Switch to using Ruby
This is a more expressive language and allows for easier reading that
some weird bash regex. I was having issues with multiline env vars etc.
so just settled on writing each one to a file. Hopefully this makes it
easier for end users to debug too by just looking at the contents of
those files if they can ssh into their app